Add FeatureKey
This generalizes and polishes up a pattern that I've seen in our internal adoption of LaunchDarkly - define a key and it's default once, to make any updates easy to manage.

It follows that the type of each key will be mostly static as well, so a way to define this once in the code is provided.